requirements.txt
Flask
gunicorn
psycopg2-binary
SQLAlchemy
flask_sqlalchemy
python-dotenv
flask-cors